@@ -1,22 +1,45 @@
-import {Navigation} from 'react-native-navigation';
-import {registerScreens} from './screens';
-registerScreens();
+import './wdyr'; // <--- must be first import
 
+import {Navigation} from 'react-native-navigation';
+import {LocaleConfig} from 'react-native-calendars';
+import {registerScreens} from './screens';
+
+registerScreens();
 // eslint-disable-next-line no-console
 console.ignoredYellowBox = ['Remote debugger'];
 
 /*
-import {LocaleConfig} from 'react-native-calendars';
-
 LocaleConfig.locales['fr'] = {
   monthNames: ['Janvier','Février','Mars','Avril','Mai','Juin','Juillet','Août','Septembre','Octobre','Novembre','Décembre'],
   monthNamesShort: ['Janv.','Févr.','Mars','Avril','Mai','Juin','Juil.','Août','Sept.','Oct.','Nov.','Déc.'],
   dayNames: ['Dimanche','Lundi','Mardi','Mercredi','Jeudi','Vendredi','Samedi'],
-  dayNamesShort: ['Dim.','Lun.','Mar.','Mer.','Jeu.','Ven.','Sam.']
+  dayNamesShort: ['Dim.','Lun.','Mar.','Mer.','Jeu.','Ven.','Sam.'],
+  today: 'Aujourd\'hui'
 };
-
 LocaleConfig.defaultLocale = 'fr';
 */
+LocaleConfig.locales['en'] = {
+  formatAccessibilityLabel: "dddd d 'of' MMMM 'of' yyyy",
+  monthNames: [
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+    'August',
+    'September',
+    'October',
+    'November',
+    'December'
+  ],
+  monthNamesShort: ['jan', 'feb', 'mar', 'apr', 'may', 'jun', 'jul', 'aug', 'sep', 'oct', 'nov', 'dec'],
+  dayNames: ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'],
+  dayNamesShort: ['S', 'M', 'T', 'W', 'T', 'F', 'S']
+};
+
+LocaleConfig.defaultLocale = 'en';
 
 Navigation.events().registerAppLaunchedListener(() => {
   Navigation.setRoot({
@@ -29,28 +52,14 @@ Navigation.events().registerAppLaunchedListener(() => {
               options: {
                 topBar: {
                   title: {
-                    text: 'WixCal',
-                  },
-                },
-              },
-            },
-          },
-        ],
-      },
-    },
+                    text: 'Wix RN Calendars'
+                  }
+                }
+              }
+            }
+          }
+        ]
+      }
+    }
   });
 });
-
-// Navigation.startSingleScreenApp({
-//   screen: {
-//     screen: 'Menu',
-//     title: 'WixCal',
-//   },
-//   appStyle: {
-//     navBarBackgroundColor: '#00adf5',
-//     navBarTextColor: 'white',
-//     navBarButtonColor: '#ffffff',
-//     statusBarTextColorScheme: 'light',
-//     autoAdjustScrollViewInsets: true
-//   }
-// });
