@@ -1,56 +1,136 @@
-import React, {Component} from 'react';
-import {
-  Text,
-  StyleSheet,
-  ScrollView,
-  View
-} from 'react-native';
+import _ from 'lodash';
+import moment from 'moment';
+import React, {useState, Fragment} from 'react';
+import {StyleSheet, View, ScrollView, Text, TouchableOpacity} from 'react-native';
 import {Calendar} from 'react-native-calendars';
 
-export default class CalendarsScreen extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-    this.onDayPress = this.onDayPress.bind(this);
-  }
+const testIDs = require('../testIDs');
 
-  render() {
+const CalendarsScreen = () => {
+  const [selected, setSelected] = useState('');
+
+  const onDayPress = day => {
+    setSelected(day.dateString);
+  };
+
+  const getDisabledDates = (startDate, endDate, daysToDisable) => {
+    const disabledDates = {};
+    const start = moment(startDate);
+    const end = moment(endDate);
+
+    for (let m = moment(start); m.diff(end, 'days') <= 0; m.add(1, 'days')) {
+      if (_.includes(daysToDisable, m.weekday())) {
+        disabledDates[m.format('YYYY-MM-DD')] = {disabled: true};
+      }
+    }
+    return disabledDates;
+  };
+
+  const renderCalendarWithSelectableDate = () => {
     return (
-      <ScrollView style={styles.container}>
-        <Text style={styles.text}>Calendar with selectable date and arrows</Text>
+      <Fragment>
+        <Text style={styles.text}>Calendar with selectable date</Text>
         <Calendar
-          onDayPress={this.onDayPress}
+          testID={testIDs.calendars.FIRST}
+          current={'2020-02-02'}
           style={styles.calendar}
-          hideExtraDays
-          markedDates={{[this.state.selected]: {selected: true, disableTouchEvent: true, selectedDotColor: 'orange'}}}
+          onDayPress={onDayPress}
+          markedDates={{
+            [selected]: {
+              selected: true,
+              disableTouchEvent: true,
+              selectedColor: 'orange',
+              selectedTextColor: 'red'
+            }
+          }}
         />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithWeekNumbers = () => {
+    return (
+      <Fragment>
+        <Text style={styles.text}>Calendar with week numbers</Text>
+        <Calendar style={styles.calendar} hideExtraDays showWeekNumbers />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithMinAndMaxDates = () => {
+    return (
+      <Fragment>
+        <Text style={styles.text}>Calendar with min and max dates</Text>
+        <Calendar 
+          style={styles.calendar} 
+          hideExtraDays 
+          current={'2012-05-16'}
+          minDate={'2012-05-10'}
+          maxDate={'2012-05-20'}
+        />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithMarkedDatesAndHiddenArrows = () => {
+    return (
+      <Fragment>
         <Text style={styles.text}>Calendar with marked dates and hidden arrows</Text>
         <Calendar
           style={styles.calendar}
           current={'2012-05-16'}
-          minDate={'2012-05-10'}
-          maxDate={'2012-05-29'}
+          hideExtraDays
+          disableAllTouchEventsForDisabledDays
           firstDay={1}
           markedDates={{
-            '2012-05-23': {selected: true, marked: true},
-            '2012-05-24': {selected: true, marked: true, dotColor: 'green'},
-            '2012-05-25': {marked: true, dotColor: 'red'},
+            '2012-05-23': {selected: true, marked: true, disableTouchEvent: true},
+            '2012-05-24': {selected: true, marked: true, dotColor: 'red'},
+            '2012-05-25': {marked: true, dotColor: 'red', disableTouchEvent: true},
             '2012-05-26': {marked: true},
-            '2012-05-27': {disabled: true, activeOpacity: 0}
+            '2012-05-27': {disabled: true, activeOpacity: 0, disableTouchEvent: false}
           }}
-          // disabledByDefault={true}
           hideArrows={true}
+          // disabledByDefault={true}
         />
-        <Text style={styles.text}>Calendar with custom day component</Text>
-        <Calendar
-          style={[styles.calendar, {height: 300}]}
-          dayComponent={({date, state}) => {
-            return (<View><Text style={{textAlign: 'center', color: state === 'disabled' ? 'gray' : 'black'}}>{date.day}</Text></View>);
-          }}
-        />
-        <Text style={styles.text}>Calendar with period marking and spinner</Text>
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithMultiDotMarking = () => {
+    return (
+      <Fragment>
+        <Text style={styles.text}>Calendar with multi-dot marking</Text>
         <Calendar
           style={styles.calendar}
+          current={'2012-05-16'}
+          markingType={'multi-dot'}
+          markedDates={{
+            '2012-05-08': {
+              selected: true,
+              dots: [
+                {key: 'vacation', color: 'blue', selectedDotColor: 'red'},
+                {key: 'massage', color: 'red', selectedDotColor: 'white'}
+              ]
+            },
+            '2012-05-09': {
+              disabled: true,
+              dots: [
+                {key: 'vacation', color: 'green', selectedDotColor: 'red'},
+                {key: 'massage', color: 'red', selectedDotColor: 'green'}
+              ]
+            }
+          }}
+        />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithPeriodMarkingAndSpinner = () => {
+    return (
+      <Fragment>
+        <Text style={styles.text}>Calendar with period marking and spinner</Text>
+        <Calendar
+          // style={styles.calendar}
           current={'2012-05-16'}
           minDate={'2012-05-10'}
           displayLoadingIndicator
@@ -58,16 +138,19 @@ export default class CalendarsScreen extends Component {
           theme={{
             calendarBackground: '#333248',
             textSectionTitleColor: 'white',
+            textSectionTitleDisabledColor: 'gray',
             dayTextColor: 'red',
             todayTextColor: 'white',
             selectedDayTextColor: 'white',
             monthTextColor: 'white',
+            indicatorColor: 'white',
             selectedDayBackgroundColor: '#333248',
             arrowColor: 'white',
             // textDisabledColor: 'red',
             'stylesheet.calendar.header': {
               week: {
-                marginTop: 5,
+                marginTop: 30,
+                marginHorizontal: 12,
                 flexDirection: 'row',
                 justifyContent: 'space-between'
               }
@@ -75,64 +158,96 @@ export default class CalendarsScreen extends Component {
           }}
           markedDates={{
             '2012-05-17': {disabled: true},
-            '2012-05-08': {textColor: '#666'},
-            '2012-05-09': {textColor: '#666'},
-            '2012-05-14': {startingDay: true, color: 'blue', endingDay: true},
-            '2012-05-21': {startingDay: true, color: 'blue'},
+            '2012-05-08': {textColor: 'pink'},
+            '2012-05-09': {textColor: 'pink'},
+            '2012-05-14': {startingDay: true, color: 'green', endingDay: true},
+            '2012-05-21': {startingDay: true, color: 'green'},
             '2012-05-22': {endingDay: true, color: 'gray'},
             '2012-05-24': {startingDay: true, color: 'gray'},
             '2012-05-25': {color: 'gray'},
-            '2012-05-26': {endingDay: true, color: 'gray'}}}
-          hideArrows={false}
-        />
-        <Text style={styles.text}>Calendar with multi-dot marking</Text>
-        <Calendar
-          style={styles.calendar}
-          current={'2012-05-16'}
-          markingType={'multi-dot'}
-          markedDates={{
-            '2012-05-08': {dots: [{key: 'vacation', color: 'blue', selectedDotColor: 'white'}, {key: 'massage', color: 'red', selectedDotColor: 'white'}], selected: true},
-            '2012-05-09': {dots: [{key: 'vacation', color: 'blue', selectedDotColor: 'red'}, {key: 'massage', color: 'red', selectedDotColor: 'blue'}], disabled: true}
+            '2012-05-26': {endingDay: true, color: 'gray'}
           }}
-          hideArrows={false}
         />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithPeriodMarkingAndDotMarking = () => {
+    return (
+      <Fragment>
+        <Text style={styles.text}>Calendar with period marking and dot marking</Text>
+        <Calendar
+          current={'2012-05-16'}
+          minDate={'2012-05-01'}
+          disabledDaysIndexes={[0, 6]}
+          markingType={'period'}
+          markedDates={{
+            '2012-05-15': {marked: true, dotColor: '#50cebb'},
+            '2012-05-16': {marked: true, dotColor: '#50cebb'},
+            '2012-05-21': {startingDay: true, color: '#50cebb', textColor: 'white'},
+            '2012-05-22': {
+              color: '#70d7c7',
+              customTextStyle: {
+                color: '#FFFAAA',
+                fontWeight: '700'
+              }
+            },
+            '2012-05-23': {color: '#70d7c7', textColor: 'white', marked: true, dotColor: 'white'},
+            '2012-05-24': {color: '#70d7c7', textColor: 'white'},
+            '2012-05-25': {
+              endingDay: true,
+              color: '#50cebb',
+              textColor: 'white',
+              customContainerStyle: {
+                borderTopRightRadius: 5,
+                borderBottomRightRadius: 5
+              }
+            },
+            ...getDisabledDates('2012-05-01', '2012-05-30', [0, 6])
+          }}
+        />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithMultiPeriodMarking = () => {
+    return (
+      <Fragment>
         <Text style={styles.text}>Calendar with multi-period marking</Text>
         <Calendar
           style={styles.calendar}
           current={'2012-05-16'}
           markingType={'multi-period'}
-          markedDates={{  
-            '2012-05-16': {  
-              periods: [  
-                { startingDay: true, endingDay: false, color: '#5f9ea0' },
-                { startingDay: true, endingDay: false, color: '#ffa500' },
+          markedDates={{
+            '2012-05-16': {
+              periods: [
+                {startingDay: true, endingDay: false, color: 'green'},
+                {startingDay: true, endingDay: false, color: 'orange'}
               ]
             },
-            '2012-05-17': {  
-              periods: [  
-                { startingDay: false, endingDay: true, color: '#5f9ea0' },
-                { startingDay: false, endingDay: true, color: '#ffa500' },
-                { startingDay: true, endingDay: false, color: '#f0e68c' },
+            '2012-05-17': {
+              periods: [
+                {startingDay: false, endingDay: true, color: 'green'},
+                {startingDay: false, endingDay: true, color: 'orange'},
+                {startingDay: true, endingDay: false, color: 'pink'}
               ]
             },
-            '2012-05-18': {  
-              periods: [  
-                { startingDay: true, endingDay: true, color: '#ffa500' },
-                { color: 'transparent' },
-                { startingDay: false, endingDay: false, color: '#f0e68c' },
+            '2012-05-18': {
+              periods: [
+                {startingDay: true, endingDay: true, color: 'orange'},
+                {color: 'transparent'},
+                {startingDay: false, endingDay: false, color: 'pink'}
               ]
-            },
+            }
           }}
-          hideArrows={false}
         />
-        <Text style={styles.text}>Calendar with week numbers</Text>
-        <Calendar
-          onDayPress={this.onDayPress}
-          style={styles.calendar}
-          hideExtraDays
-          showWeekNumbers
-          markedDates={{[this.state.selected]: {selected: true}}}
-        />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithCustomMarkingType = () => {
+    return (
+      <Fragment>
         <Text style={styles.text}>Custom calendar with custom marking type</Text>
         <Calendar
           style={styles.calendar}
@@ -149,32 +264,33 @@ export default class CalendarsScreen extends Component {
                   elevation: 2
                 },
                 text: {
-                  color: 'blue',
-                },
+                  color: 'red'
+                }
               }
             },
-            '2018-03-08': {selected: true},
+            '2018-03-08': {
+              selected: true
+            },
             '2018-03-09': {
               customStyles: {
                 container: {
                   backgroundColor: 'red',
-                  elevation: 4,
+                  elevation: 4
                 },
                 text: {
-                  color: 'white',
-                },
+                  color: 'white'
+                }
               }
             },
-            '2018-03-10': {disabled: true},
             '2018-03-14': {
               customStyles: {
                 container: {
-                  backgroundColor: 'green',
+                  backgroundColor: 'green'
                 },
                 text: {
-                  color: 'white',
-                },
-              },
+                  color: 'white'
+                }
+              }
             },
             '2018-03-15': {
               customStyles: {
@@ -183,99 +299,142 @@ export default class CalendarsScreen extends Component {
                   elevation: 2
                 },
                 text: {
-                  color: 'yellow',
-                },
+                  color: 'yellow'
+                }
               }
             },
-            '2018-03-20': {
-              customStyles: {
-                container: {
-                  backgroundColor: 'pink',
-                  elevation: 4,
-                },
-                text: {
-                  color: 'blue',
-                },
-              }
+            '2018-03-21': {
+              disabled: true
             },
-            '2018-03-21': {disabled: true},
             '2018-03-28': {
               customStyles: {
-                container: {
-                  backgroundColor: 'green',
-                },
                 text: {
                   color: 'black',
                   fontWeight: 'bold'
-                },
-              },
-            },
-            '2018-03-29': {
-              customStyles: {
-                container: {
-                  backgroundColor: 'white',
-                  elevation: 2
-                },
-                text: {
-                  color: 'blue',
-                },
+                }
               }
             },
             '2018-03-30': {
               customStyles: {
                 container: {
-                  backgroundColor: 'violet',
+                  backgroundColor: 'pink',
                   elevation: 4,
-                  borderColor: 'red',
-                  borderWidth: 5,
+                  borderColor: 'purple',
+                  borderWidth: 5
                 },
                 text: {
                   marginTop: 3,
                   fontSize: 11,
-                  color: 'yellow',
-                },
+                  color: 'black'
+                }
               }
             },
             '2018-03-31': {
               customStyles: {
                 container: {
-                  backgroundColor: 'green',
-                  borderRadius: 0,
-                },
-                text: {
-                  color: 'white',
-                },
-              },
-            }}}
-          hideArrows={false}
+                  backgroundColor: 'orange',
+                  borderRadius: 0
+                }
+              }
+            }
+          }}
         />
-      </ScrollView>
+      </Fragment>
     );
-  }
+  };
 
-  onDayPress(day) {
-    this.setState({
-      selected: day.dateString
+  const renderCalendarWithCustomDay = () => {
+    return (
+      <Fragment>
+        <Text style={styles.text}>Calendar with custom day component</Text>
+        <Calendar
+          style={[
+            styles.calendar,
+            {
+              height: 250,
+              borderBottomWidth: 1,
+              borderBottomColor: 'lightgrey'
+            }
+          ]}
+          dayComponent={({date, state}) => {
+            return (
+              <View>
+                <Text style={{textAlign: 'center', color: state === 'disabled' ? 'gray' : 'purple'}}>{date.day}</Text>
+              </View>
+            );
+          }}
+        />
+      </Fragment>
+    );
+  };
+
+  const renderCalendarWithCustomHeader = () => {
+    const CustomHeader = React.forwardRef((props, ref) => {
+      return (
+        <View
+          ref={ref}
+          {...props}
+          style={{
+            backgroundColor: '#FCC',
+            flexDirection: 'row',
+            justifyContent: 'space-around',
+            marginHorizontal: -4,
+            padding: 8
+          }}
+        >
+          <Text>This is a custom header!</Text>
+          <TouchableOpacity onPress={() => console.warn('Tapped!')}>
+            <Text>Tap Me</Text>
+          </TouchableOpacity>
+        </View>
+      );
     });
-  }
-}
+
+    return (
+      <Fragment>
+        <Text style={styles.text}>Calendar with custom header component</Text>
+        <Calendar
+          testID={testIDs.calendars.LAST}
+          style={[
+            styles.calendar,
+            {
+              height: 250,
+              borderBottomWidth: 1,
+              borderBottomColor: 'lightgrey'
+            }
+          ]}
+          customHeader={CustomHeader}
+        />
+      </Fragment>
+    );
+  };
+  return (
+    <ScrollView showsVerticalScrollIndicator={false} testID={testIDs.calendars.CONTAINER}>
+      {renderCalendarWithSelectableDate()}
+      {renderCalendarWithWeekNumbers()}
+      {renderCalendarWithMinAndMaxDates()}
+      {renderCalendarWithMarkedDatesAndHiddenArrows()}
+      {renderCalendarWithMultiDotMarking()}
+      {renderCalendarWithPeriodMarkingAndSpinner()}
+      {renderCalendarWithPeriodMarkingAndDotMarking()}
+      {renderCalendarWithMultiPeriodMarking()}
+      {renderCalendarWithCustomMarkingType()}
+      {renderCalendarWithCustomDay()}
+      {renderCalendarWithCustomHeader()}
+    </ScrollView>
+  );
+};
+
+export default CalendarsScreen;
 
 const styles = StyleSheet.create({
   calendar: {
-    borderTopWidth: 1,
-    paddingTop: 5,
-    borderBottomWidth: 1,
-    borderColor: '#eee',
-    height: 350
+    marginBottom: 10
   },
   text: {
     textAlign: 'center',
-    borderColor: '#bbb',
     padding: 10,
-    backgroundColor: '#eee'
-  },
-  container: {
-    flex: 1,
-    backgroundColor: 'gray'
+    backgroundColor: 'lightgrey',
+    fontSize: 16
   }
 });
