@@ -1,25 +1,36 @@
-import React, {Component} from 'react';
-
+import React from 'react';
 import {CalendarList} from 'react-native-calendars';
-import {View} from 'react-native';
 
-export default class HorizontalCalendarList extends Component {
-  constructor(props) {
-    super(props);
-  }
+const testIDs = require('../testIDs');
 
-  render() {
-    return (
-      <View>
-        <CalendarList
-          current={'2012-05-16'}
-          pastScrollRange={24}
-          futureScrollRange={24}
-          horizontal
-          pagingEnabled
-          style={{borderBottomWidth: 1, borderBottomColor: 'black'}}
-        />
-      </View>
-    );
-  }
-}
+const HorizontalCalendarList = () => {
+  const [selectedDate, setSelectedDate] = React.useState('2020-05-16');
+  const [markedDates, setMarkedDates] = React.useState({});
+
+  const setNewDaySelected = date => {
+    const markedDate = Object.assign({});
+    markedDate[date] = {
+      selected: true,
+      selectedColor: '#DFA460'
+    };
+    setSelectedDate(date);
+    setMarkedDates(markedDate);
+  };
+
+  return (
+    <CalendarList
+      testID={testIDs.horizontalList.CONTAINER}
+      markedDates={markedDates}
+      current={selectedDate}
+      pastScrollRange={24}
+      futureScrollRange={24}
+      horizontal
+      pagingEnabled
+      onDayPress={day => {
+        setNewDaySelected(day.dateString);
+      }}
+    />
+  );
+};
+
+export default HorizontalCalendarList;
