@@ -6,21 +6,17 @@ const STYLESHEET_ID = 'stylesheet.agenda.main';
 
 export default function styleConstructor(theme = {}) {
   const appStyle = {...defaultStyle, ...theme};
-  const { knob, weekdays } = platformStyles(appStyle);
+  const {knob, weekdays} = platformStyles(appStyle);
+
   return StyleSheet.create({
     knob,
     weekdays,
     header: {
       overflow: 'hidden',
       justifyContent: 'flex-end',
-      position:'absolute',
-      height:'100%',
-      width:'100%',
-    },
-    calendar: {
-      flex: 1,
-      borderBottomWidth: 1,
-      borderColor: appStyle.separatorColor
+      position: 'absolute',
+      height: '100%',
+      width: '100%'
     },
     knobContainer: {
       flex: 1,
@@ -35,8 +31,10 @@ export default function styleConstructor(theme = {}) {
     weekday: {
       width: 32,
       textAlign: 'center',
-      fontSize: 13,
       color: appStyle.textSectionTitleColor,
+      fontSize: appStyle.textDayHeaderFontSize,
+      fontFamily: appStyle.textDayHeaderFontFamily,
+      fontWeight: appStyle.textDayHeaderFontWeight
     },
     reservations: {
       flex: 1,
