@@ -1,80 +1,102 @@
-import React, {Component} from 'react';
-import {
-  FlatList, Platform, Dimensions,
-} from 'react-native';
+import _ from 'lodash';
 import PropTypes from 'prop-types';
 import XDate from 'xdate';
-
+import React, {Component} from 'react';
+import {FlatList, Platform, Dimensions, View} from 'react-native';
+import {extractComponentProps} from '../component-updater';
 import {xdateToData, parseDate} from '../interface';
-import styleConstructor from './style';
 import dateutils from '../dateutils';
+import {STATIC_HEADER} from '../testIDs';
+import styleConstructor from './style';
 import Calendar from '../calendar';
 import CalendarListItem from './item';
+import CalendarHeader from '../calendar/header/index';
 
 const {width} = Dimensions.get('window');
 
+/**
+ * @description: Calendar List component for both vertical and horizontal calendars
+ * @extends: Calendar
+ * @extendslink: docs/Calendar
+ * @example: https://github.com/wix/react-native-calendars/blob/master/example/src/screens/calendarsList.js
+ * @gif: https://github.com/wix/react-native-calendars/blob/master/demo/calendar-list.gif
+ */
 class CalendarList extends Component {
+  static displayName = 'CalendarList';
+
   static propTypes = {
     ...Calendar.propTypes,
-
-    // Max amount of months allowed to scroll to the past. Default = 50
+    /** Max amount of months allowed to scroll to the past. Default = 50 */
     pastScrollRange: PropTypes.number,
-
-    // Max amount of months allowed to scroll to the future. Default = 50
+    /** Max amount of months allowed to scroll to the future. Default = 50 */
     futureScrollRange: PropTypes.number,
-
-    // Enable or disable scrolling of calendar list
-    scrollEnabled: PropTypes.bool,
-
-    // Enable or disable vertical scroll indicator. Default = false
+    /** Used when calendar scroll is horizontal, default is device width, pagination should be disabled */
+    calendarWidth: PropTypes.number,
+    /** Dynamic calendar height */
+    calendarHeight: PropTypes.number,
+    /** Style for the List item (the calendar) */
+    calendarStyle: PropTypes.oneOfType([PropTypes.object, PropTypes.number, PropTypes.array]),
+    /** Whether to use static header that will not scroll with the list (horizontal only) */
+    staticHeader: PropTypes.bool,
+    /** Enable or disable vertical / horizontal scroll indicator. Default = false */
     showScrollIndicator: PropTypes.bool,
 
-    // When true, the calendar list scrolls to top when the status bar is tapped. Default = true
+    /** Enable or disable scrolling of calendar list */
+    scrollEnabled: PropTypes.bool,
+    /** When true, the calendar list scrolls to top when the status bar is tapped. Default = true */
     scrollsToTop: PropTypes.bool,
-
-    // Enable or disable paging on scroll
+    /** Enable or disable paging on scroll */
     pagingEnabled: PropTypes.bool,
-
-    // Used when calendar scroll is horizontal, default is device width, pagination should be disabled
-    calendarWidth: PropTypes.number,
-
-    // Whether the scroll is horizontal
+    /** Whether the scroll is horizontal */
     horizontal: PropTypes.bool,
-    // Dynamic calendar height
-    calendarHeight: PropTypes.number,
-
-    // Style for the List item (the calendar)
-    calendarStyle: PropTypes.oneOfType([PropTypes.object, PropTypes.number])
-  }
+    /** Should Keyboard persist taps */
+    keyboardShouldPersistTaps: PropTypes.oneOf(['never', 'always', 'handled']),
+    /** A custom key extractor for the generated calendar months */
+    keyExtractor: PropTypes.func,
+    /** How far from the end to trigger the onEndReached callback */
+    onEndReachedThreshold: PropTypes.number,
+    /** Called once when the scroll position gets within onEndReachedThreshold */
+    onEndReached: PropTypes.func
+  };
 
   static defaultProps = {
-    horizontal: false,
     calendarWidth: width,
     calendarHeight: 360,
     pastScrollRange: 50,
     futureScrollRange: 50,
     showScrollIndicator: false,
-    scrollEnabled: true,
+    horizontal: false,
     scrollsToTop: false,
-    removeClippedSubviews: Platform.OS === 'android' ? false : true,
-  }
+    scrollEnabled: true,
+    removeClippedSubviews: Platform.OS === 'android',
+    keyExtractor: (item, index) => String(index)
+  };
 
   constructor(props) {
     super(props);
+
     this.style = styleConstructor(props.theme);
+
+    this.viewabilityConfig = {
+      itemVisiblePercentThreshold: 20
+    };
 
     const rows = [];
     const texts = [];
     const date = parseDate(props.current) || XDate();
-    for (let i = 0; i <= this.props.pastScrollRange + this.props.futureScrollRange; i++) {
-      const rangeDate = date.clone().addMonths(i - this.props.pastScrollRange, true);
+
+    for (let i = 0; i <= props.pastScrollRange + props.futureScrollRange; i++) {
+      const rangeDate = date.clone().addMonths(i - props.pastScrollRange, true);
       const rangeDateStr = rangeDate.toString('MMM yyyy');
       texts.push(rangeDateStr);
       /*
        * This selects range around current shown month [-0, +2] or [-1, +1] month for detail calendar rendering.
        * If `this.pastScrollRange` is `undefined` it's equal to `false` or 0 in next condition.
        */
-      if (this.props.pastScrollRange - 1 <= i && i <= this.props.pastScrollRange + 1 || !this.props.pastScrollRange && i <= this.props.pastScrollRange + 2) {
+      if (
+        (props.pastScrollRange - 1 <= i && i <= props.pastScrollRange + 1) ||
+        (!props.pastScrollRange && i <= props.pastScrollRange + 2)
+      ) {
         rows.push(rangeDate);
       } else {
         rows.push(rangeDateStr);
@@ -84,29 +106,45 @@ class CalendarList extends Component {
     this.state = {
       rows,
       texts,
-      openDate: date
+      openDate: date,
+      currentMonth: parseDate(props.current)
     };
-
-    this.onViewableItemsChangedBound = this.onViewableItemsChanged.bind(this);
-    this.renderCalendarBound = this.renderCalendar.bind(this);
-    this.getItemLayout = this.getItemLayout.bind(this);
-    this.onLayout = this.onLayout.bind(this);
   }
 
-  onLayout(event) {
-    if (this.props.onLayout) {
-      this.props.onLayout(event);
+  componentDidUpdate(prevProps) {
+    const prevCurrent = parseDate(prevProps.current);
+    const current = parseDate(this.props.current);
+
+    if (current && prevCurrent && current.getTime() !== prevCurrent.getTime()) {
+      this.scrollToMonth(current);
     }
   }
 
+  static getDerivedStateFromProps(nextProps, prevState) {
+    const rowclone = prevState.rows;
+    const newrows = [];
+
+    for (let i = 0; i < rowclone.length; i++) {
+      let val = prevState.texts[i];
+      if (rowclone[i].getTime) {
+        val = rowclone[i].clone();
+        val.propbump = rowclone[i].propbump ? rowclone[i].propbump + 1 : 1;
+      }
+      newrows.push(val);
+    }
+    return {rows: newrows};
+  }
+
   scrollToDay(d, offset, animated) {
+    const {horizontal, calendarHeight, calendarWidth, pastScrollRange, firstDay} = this.props;
     const day = parseDate(d);
     const diffMonths = Math.round(this.state.openDate.clone().setDate(1).diffMonths(day.clone().setDate(1)));
-    const size = this.props.horizontal ? this.props.calendarWidth : this.props.calendarHeight;
-    let scrollAmount = (size * this.props.pastScrollRange) + (diffMonths * size) + (offset || 0);
-    if (!this.props.horizontal) {
+    const size = horizontal ? calendarWidth : calendarHeight;
+    let scrollAmount = size * pastScrollRange + diffMonths * size + (offset || 0);
+
+    if (!horizontal) {
       let week = 0;
-      const days = dateutils.page(day, this.props.firstDay);
+      const days = dateutils.page(day, firstDay);
       for (let i = 0; i < days.length; i++) {
         week = Math.floor(i / 7);
         if (dateutils.sameDate(days[i], day)) {
@@ -118,40 +156,59 @@ class CalendarList extends Component {
     this.listView.scrollToOffset({offset: scrollAmount, animated});
   }
 
-  scrollToMonth(m) {
+  scrollToMonth = m => {
+    const {horizontal, calendarHeight, calendarWidth, pastScrollRange} = this.props;
     const month = parseDate(m);
     const scrollTo = month || this.state.openDate;
     let diffMonths = Math.round(this.state.openDate.clone().setDate(1).diffMonths(scrollTo.clone().setDate(1)));
-    const size = this.props.horizontal ? this.props.calendarWidth : this.props.calendarHeight;
-    const scrollAmount = (size * this.props.pastScrollRange) + (diffMonths * size);
-    //console.log(month, this.state.openDate);
-    //console.log(scrollAmount, diffMonths);
+    const size = horizontal ? calendarWidth : calendarHeight;
+    const scrollAmount = size * pastScrollRange + diffMonths * size;
+
     this.listView.scrollToOffset({offset: scrollAmount, animated: false});
+  };
+
+  getItemLayout = (data, index) => {
+    const {horizontal, calendarHeight, calendarWidth} = this.props;
+
+    return {
+      length: horizontal ? calendarWidth : calendarHeight,
+      offset: (horizontal ? calendarWidth : calendarHeight) * index,
+      index
+    };
+  };
+
+  getMonthIndex(month) {
+    let diffMonths = this.state.openDate.diffMonths(month) + this.props.pastScrollRange;
+    return diffMonths;
   }
 
-  componentWillReceiveProps(props) {
-    const current = parseDate(this.props.current);
-    const nextCurrent = parseDate(props.current);
-    if (nextCurrent && current && nextCurrent.getTime() !== current.getTime()) {
-      this.scrollToMonth(nextCurrent);
+  addMonth = count => {
+    this.updateMonth(this.state.currentMonth.clone().addMonths(count, true));
+  };
+
+  updateMonth(day, doNotTriggerListeners) {
+    if (day.toString('yyyy MM') === this.state.currentMonth.toString('yyyy MM')) {
+      return;
     }
 
-    const rowclone = this.state.rows;
-    const newrows = [];
-    for (let i = 0; i < rowclone.length; i++) {
-      let val = this.state.texts[i];
-      if (rowclone[i].getTime) {
-        val = rowclone[i].clone();
-        val.propbump = rowclone[i].propbump ? rowclone[i].propbump + 1 : 1;
+    this.setState(
+      {
+        currentMonth: day.clone()
+      },
+      () => {
+        this.scrollToMonth(this.state.currentMonth);
+
+        if (!doNotTriggerListeners) {
+          const currMont = this.state.currentMonth.clone();
+
+          _.invoke(this.props, 'onMonthChange', xdateToData(currMont));
+          _.invoke(this.props, 'onVisibleMonthsChange', [xdateToData(currMont)]);
+        }
       }
-      newrows.push(val);
-    }
-    this.setState({
-      rows: newrows
-    });
+    );
   }
 
-  onViewableItemsChanged({viewableItems}) {
+  onViewableItemsChanged = ({viewableItems}) => {
     function rowIsCloseToViewable(index, distance) {
       for (let i = 0; i < viewableItems.length; i++) {
         if (Math.abs(index - parseInt(viewableItems[i].index)) <= distance) {
@@ -164,9 +221,11 @@ class CalendarList extends Component {
     const rowclone = this.state.rows;
     const newrows = [];
     const visibleMonths = [];
+
     for (let i = 0; i < rowclone.length; i++) {
       let val = rowclone[i];
       const rowShouldBeRendered = rowIsCloseToViewable(i, 1);
+
       if (rowShouldBeRendered && !rowclone[i].getTime) {
         val = this.state.openDate.clone().addMonths(i - this.props.pastScrollRange, true);
       } else if (!rowShouldBeRendered) {
@@ -177,60 +236,81 @@ class CalendarList extends Component {
         visibleMonths.push(xdateToData(val));
       }
     }
-    if (this.props.onVisibleMonthsChange) {
-      this.props.onVisibleMonthsChange(visibleMonths);
-    }
-    this.setState({
-      rows: newrows
-    });
-  }
 
-  renderCalendar({item}) {
+    _.invoke(this.props, 'onVisibleMonthsChange', visibleMonths);
+
+    this.setState({
+      rows: newrows,
+      currentMonth: parseDate(visibleMonths[0])
+    });
+  };
+
+  renderItem = ({item}) => {
+    const {calendarStyle, horizontal, calendarWidth, testID, ...others} = this.props;
+
     return (
-      <CalendarListItem 
-        item={item} 
-        calendarHeight={this.props.calendarHeight} 
-        calendarWidth={this.props.horizontal ? this.props.calendarWidth : undefined} 
-        {...this.props} 
-        style={this.props.calendarStyle}
+      <CalendarListItem
+        {...others}
+        item={item}
+        testID={`${testID}_${item}`}
+        style={calendarStyle}
+        calendarWidth={horizontal ? calendarWidth : undefined}
+        scrollToMonth={this.scrollToMonth}
       />
     );
-  }
+  };
 
-  getItemLayout(data, index) {
-    return {length: this.props.horizontal ? this.props.calendarWidth : this.props.calendarHeight, offset: (this.props.horizontal ? this.props.calendarWidth : this.props.calendarHeight) * index, index};
-  }
+  renderStaticHeader() {
+    const {staticHeader, horizontal, headerStyle} = this.props;
+    const useStaticHeader = staticHeader && horizontal;
+    const headerProps = extractComponentProps(CalendarHeader, this.props);
 
-  getMonthIndex(month) {
-    let diffMonths = this.state.openDate.diffMonths(month) + this.props.pastScrollRange;
-    return diffMonths;
+    if (useStaticHeader) {
+      return (
+        <CalendarHeader
+          {...headerProps}
+          testID={STATIC_HEADER}
+          style={[this.style.staticHeader, headerStyle]}
+          month={this.state.currentMonth}
+          addMonth={this.addMonth}
+          accessibilityElementsHidden={true} // iOS
+          importantForAccessibility={'no-hide-descendants'} // Android
+        />
+      );
+    }
   }
 
   render() {
+    const {style, pastScrollRange, futureScrollRange, horizontal, showScrollIndicator, testID} = this.props;
+
     return (
-      <FlatList
-        onLayout={this.onLayout}
-        ref={(c) => this.listView = c}
-        //scrollEventThrottle={1000}
-        style={[this.style.container, this.props.style]}
-        initialListSize={this.props.pastScrollRange + this.props.futureScrollRange + 1}
-        data={this.state.rows}
-        //snapToAlignment='start'
-        //snapToInterval={this.calendarHeight}
-        removeClippedSubviews={this.props.removeClippedSubviews}
-        pageSize={1}
-        horizontal={this.props.horizontal}
-        pagingEnabled={this.props.pagingEnabled}
-        onViewableItemsChanged={this.onViewableItemsChangedBound}
-        renderItem={this.renderCalendarBound}
-        showsVerticalScrollIndicator={this.props.showScrollIndicator}
-        showsHorizontalScrollIndicator={this.props.showScrollIndicator}
-        scrollEnabled={this.props.scrollEnabled}
-        keyExtractor={(item, index) => String(index)}
-        initialScrollIndex={this.state.openDate ? this.getMonthIndex(this.state.openDate) : false}
-        getItemLayout={this.getItemLayout}
-        scrollsToTop={this.props.scrollsToTop}
-      />
+      <View>
+        <FlatList
+          ref={c => (this.listView = c)}
+          style={[this.style.container, style]}
+          initialListSize={pastScrollRange + futureScrollRange + 1} // ListView deprecated
+          data={this.state.rows}
+          renderItem={this.renderItem}
+          getItemLayout={this.getItemLayout}
+          onViewableItemsChanged={this.onViewableItemsChanged}
+          viewabilityConfig={this.viewabilityConfig}
+          initialScrollIndex={this.state.openDate ? this.getMonthIndex(this.state.openDate) : false}
+          showsVerticalScrollIndicator={showScrollIndicator}
+          showsHorizontalScrollIndicator={horizontal && showScrollIndicator}
+          testID={testID}
+          onLayout={this.props.onLayout}
+          removeClippedSubviews={this.props.removeClippedSubviews}
+          pagingEnabled={this.props.pagingEnabled}
+          scrollEnabled={this.props.scrollEnabled}
+          scrollsToTop={this.props.scrollsToTop}
+          horizontal={this.props.horizontal}
+          keyboardShouldPersistTaps={this.props.keyboardShouldPersistTaps}
+          keyExtractor={this.props.keyExtractor}
+          onEndReachedThreshold={this.props.onEndReachedThreshold}
+          onEndReached={this.props.onEndReached}
+        />
+        {this.renderStaticHeader()}
+      </View>
     );
   }
 }
