@@ -3,9 +3,13 @@ import * as defaultStyle from '../../../style';
 
 const STYLESHEET_ID = 'stylesheet.day.basic';
 
-export default function styleConstructor(theme={}) {
+export default function styleConstructor(theme = {}) {
   const appStyle = {...defaultStyle, ...theme};
   return StyleSheet.create({
+    container: {
+      alignSelf: 'stretch',
+      alignItems: 'center'
+    },
     base: {
       width: 32,
       height: 32,
@@ -15,9 +19,10 @@ export default function styleConstructor(theme={}) {
       marginTop: Platform.OS === 'android' ? 4 : 6,
       fontSize: appStyle.textDayFontSize,
       fontFamily: appStyle.textDayFontFamily,
-      fontWeight: '300',
+      fontWeight: appStyle.textDayFontWeight,
       color: appStyle.dayTextColor,
-      backgroundColor: 'rgba(255, 255, 255, 0)'
+      backgroundColor: 'rgba(255, 255, 255, 0)',
+      ...appStyle.textDayStyle
     },
     alignedText: {
       marginTop: Platform.OS === 'android' ? 4 : 6
@@ -27,7 +32,8 @@ export default function styleConstructor(theme={}) {
       borderRadius: 16
     },
     today: {
-      backgroundColor: appStyle.todayBackgroundColor
+      backgroundColor: appStyle.todayBackgroundColor,
+      borderRadius: 16
     },
     todayText: {
       color: appStyle.todayTextColor
@@ -43,7 +49,8 @@ export default function styleConstructor(theme={}) {
       height: 4,
       marginTop: 1,
       borderRadius: 2,
-      opacity: 0
+      opacity: 0,
+      ...appStyle.dotStyle
     },
     visibleDot: {
       opacity: 1,
@@ -51,6 +58,12 @@ export default function styleConstructor(theme={}) {
     },
     selectedDot: {
       backgroundColor: appStyle.selectedDotColor
+    },
+    disabledDot: {
+      backgroundColor: appStyle.disabledDotColor || appStyle.dotColor
+    },
+    todayDot: {
+      backgroundColor: appStyle.todayDotColor || appStyle.dotColor
     },
     ...(theme[STYLESHEET_ID] || {})
   });
