@@ -1,71 +1,68 @@
-import React, {Component} from 'react';
+import _ from 'lodash';
 import PropTypes from 'prop-types';
-import {
-  TouchableWithoutFeedback,
-  Text,
-  View} from 'react-native';
+import React, {Component} from 'react';
+import {TouchableWithoutFeedback, Text, View} from 'react-native';
 import {shouldUpdate} from '../../../component-updater';
-import isEqual from 'lodash.isequal';
-
 import * as defaultStyle from '../../../style';
 import styleConstructor from './style';
+import Dot from '../dot';
 
-class Day extends Component {
+
+export default class PeriodDay extends Component {
+  static displayName = 'IGNORE';
+
   static propTypes = {
-    // TODO: selected + disabled props should be removed
-    state: PropTypes.oneOf(['selected', 'disabled', 'today', '']),
-
-    // Specify theme properties to override specific styles for calendar parts. Default = {}
-    theme: PropTypes.object,
+    state: PropTypes.oneOf(['selected', 'disabled', 'today', '']), //TODO: deprecate
     marking: PropTypes.any,
-
+    theme: PropTypes.object,
     onPress: PropTypes.func,
     onLongPress: PropTypes.func,
-    date: PropTypes.object,
-
-    markingExists: PropTypes.bool,
+    date: PropTypes.object
   };
 
   constructor(props) {
     super(props);
+
     this.theme = {...defaultStyle, ...(props.theme || {})};
     this.style = styleConstructor(props.theme);
+    
     this.markingStyle = this.getDrawingStyle(props.marking || []);
-    this.onDayPress = this.onDayPress.bind(this);
-    this.onDayLongPress = this.onDayLongPress.bind(this);
   }
 
-  onDayPress() {
+  onPress = () => {
     this.props.onPress(this.props.date);
   }
 
-  onDayLongPress() {
+  onLongPress = () => {
     this.props.onLongPress(this.props.date);
   }
 
   shouldComponentUpdate(nextProps) {
     const newMarkingStyle = this.getDrawingStyle(nextProps.marking);
-
-    if (!isEqual(this.markingStyle, newMarkingStyle)) {
+    if (!_.isEqual(this.markingStyle, newMarkingStyle)) {
       this.markingStyle = newMarkingStyle;
       return true;
     }
 
-    return shouldUpdate(this.props, nextProps, ['state', 'children', 'onPress', 'onLongPress']);
+    return shouldUpdate(this.props, nextProps, ['children', 'state', 'marking', 'onPress', 'onLongPress', 'date']);
   }
 
   getDrawingStyle(marking) {
-    const defaultStyle = {textStyle: {}};
+    const defaultStyle = {textStyle: {}, containerStyle: {}};
+    
     if (!marking) {
       return defaultStyle;
     }
+
     if (marking.disabled) {
-      defaultStyle.textStyle.color = this.theme.textDisabledColor;
+      defaultStyle.textStyle.color = this.style.disabledText.color;
     } else if (marking.selected) {
-      defaultStyle.textStyle.color = this.theme.selectedDayTextColor;
+      defaultStyle.textStyle.color = this.style.selectedText.color;
     }
+
     const resultStyle = ([marking]).reduce((prev, next) => {
-      if (next.quickAction) {
+      
+      if (next.quickAction) { //???
         if (next.first || next.last) {
           prev.containerStyle = this.style.firstQuickAction;
           prev.textStyle = this.style.firstQuickActionText;
@@ -84,27 +81,28 @@ class Day extends Component {
         return prev;
       }
 
-      const color = next.color;
-      if (next.status === 'NotAvailable') {
+      if (next.status === 'NotAvailable') { //???
         prev.textStyle = this.style.naText;
       }
+      
+      const color = next.color;
       if (next.startingDay) {
-        prev.startingDay = {
-          color
-        };
+        prev.startingDay = {color};
       }
       if (next.endingDay) {
-        prev.endingDay = {
-          color
-        };
+        prev.endingDay = {color};
       }
       if (!next.startingDay && !next.endingDay) {
-        prev.day = {
-          color
-        };
+        prev.day = {color};
       }
       if (next.textColor) {
         prev.textStyle.color = next.textColor;
+      }
+      if (marking.customTextStyle) {
+        defaultStyle.textStyle = marking.customTextStyle;
+      }
+      if (marking.customContainerStyle) {
+        defaultStyle.containerStyle = marking.customContainerStyle;
       }
       return prev;
     }, defaultStyle);
@@ -112,6 +110,7 @@ class Day extends Component {
   }
 
   render() {
+    const {state, marking} = this.props;
     const containerStyle = [this.style.base];
     const textStyle = [this.style.text];
     let leftFillerStyle = {};
@@ -119,14 +118,14 @@ class Day extends Component {
     let fillerStyle = {};
     let fillers;
 
-    if (this.props.state === 'disabled') {
+    if (state === 'disabled') {
       textStyle.push(this.style.disabledText);
-    } else if (this.props.state === 'today') {
+    } else if (state === 'today') {
       containerStyle.push(this.style.today);
       textStyle.push(this.style.todayText);
     }
 
-    if (this.props.marking) {
+    if (marking) {
       containerStyle.push({
         borderRadius: 17
       });
@@ -168,7 +167,6 @@ class Day extends Component {
       } else if (flags.day) {
         leftFillerStyle = {backgroundColor: flags.day.color};
         rightFillerStyle = {backgroundColor: flags.day.color};
-        // #177 bug
         fillerStyle = {backgroundColor: flags.day.color};
       } else if (flags.endingDay && flags.startingDay) {
         rightFillerStyle = {
@@ -190,20 +188,30 @@ class Day extends Component {
       );
     }
 
+    const {marking: {marked, dotColor, disableTouchEvent}, theme, accessibilityLabel, testID} = this.props;
+
     return (
       <TouchableWithoutFeedback
-        testID={this.props.testID}
-        onPress={this.onDayPress}
-        onLongPress={this.onDayLongPress}>
+        testID={testID}
+        onPress={this.onPress}
+        onLongPress={this.onLongPress}
+        disabled={disableTouchEvent}
+        accessible
+        accessibilityRole={disableTouchEvent ? undefined : 'button'}
+        accessibilityLabel={accessibilityLabel}
+      >
         <View style={this.style.wrapper}>
           {fillers}
           <View style={containerStyle}>
             <Text allowFontScaling={false} style={textStyle}>{String(this.props.children)}</Text>
+            <Dot
+              theme={theme}
+              color={dotColor}
+              marked={marked}
+            />
           </View>
         </View>
       </TouchableWithoutFeedback>
     );
   }
 }
-
-export default Day;
