@@ -1,8 +1,8 @@
 import {StyleSheet} from 'react-native';
 import * as defaultStyle from '../../../style';
 
-const STYLESHEET_ID = 'stylesheet.day.period';
 
+const STYLESHEET_ID = 'stylesheet.day.period';
 const FILLER_HEIGHT = 34;
 
 export default function styleConstructor(theme={}) {
@@ -14,11 +14,11 @@ export default function styleConstructor(theme={}) {
       marginLeft: -1
     },
     base: {
-      //borderWidth: 1,
       width: 38,
       height: FILLER_HEIGHT,
       alignItems: 'center'
     },
+
     fillers: {
       position: 'absolute',
       height: FILLER_HEIGHT,
@@ -34,12 +34,13 @@ export default function styleConstructor(theme={}) {
       height: FILLER_HEIGHT,
       flex: 1
     },
+
     text: {
       marginTop: 7,
       fontSize: appStyle.textDayFontSize,
       fontFamily: appStyle.textDayFontFamily,
-      fontWeight: '300',
-      color: appStyle.dayTextColor || '#2d4150',
+      fontWeight: appStyle.textDayFontWeight,
+      color: appStyle.dayTextColor,
       backgroundColor: 'rgba(255, 255, 255, 0)'
     },
     today: {
@@ -47,12 +48,15 @@ export default function styleConstructor(theme={}) {
     },
     todayText: {
       fontWeight: '500',
-      color: theme.todayTextColor || appStyle.dayTextColor,
-      //color: appStyle.textLinkColor
+      color: theme.todayTextColor || appStyle.dayTextColor
+    },
+    selectedText: {
+      color: appStyle.selectedDayTextColor
     },
     disabledText: {
       color: appStyle.textDisabledColor
     },
+    
     quickAction: {
       backgroundColor: 'white',
       borderWidth: 1,
