@@ -3,7 +3,7 @@ import * as defaultStyle from '../../style';
 
 const STYLESHEET_ID = 'stylesheet.calendar.header';
 
-export default function(theme={}) {
+export default function (theme = {}) {
   const appStyle = {...defaultStyle, ...theme};
   return StyleSheet.create({
     header: {
@@ -11,7 +11,11 @@ export default function(theme={}) {
       justifyContent: 'space-between',
       paddingLeft: 10,
       paddingRight: 10,
+      marginTop: 6,
       alignItems: 'center'
+    },
+    headerContainer: {
+      flexDirection: 'row'
     },
     monthText: {
       fontSize: appStyle.textMonthFontSize,
@@ -21,17 +25,20 @@ export default function(theme={}) {
       margin: 10
     },
     arrow: {
-      padding: 10
+      padding: 10,
+      ...appStyle.arrowStyle
     },
     arrowImage: {
+      tintColor: appStyle.arrowColor,
       ...Platform.select({
-        ios: {
-          tintColor: appStyle.arrowColor
-        },
-        android: {
-          tintColor: appStyle.arrowColor
+        web: {
+          width: appStyle.arrowWidth,
+          height: appStyle.arrowHeight
         }
       })
+    },
+    disabledArrowImage: {
+      tintColor: appStyle.disabledArrowColor
     },
     week: {
       marginTop: 7,
@@ -45,7 +52,11 @@ export default function(theme={}) {
       textAlign: 'center',
       fontSize: appStyle.textDayHeaderFontSize,
       fontFamily: appStyle.textDayHeaderFontFamily,
+      fontWeight: appStyle.textDayHeaderFontWeight,
       color: appStyle.textSectionTitleColor
+    },
+    disabledDayHeader: {
+      color: appStyle.textSectionTitleDisabledColor
     },
     ...(theme[STYLESHEET_ID] || {})
   });
