@@ -3,13 +3,20 @@ import * as defaultStyle from '../style';
 
 const STYLESHEET_ID = 'stylesheet.calendar.main';
 
-export default function getStyle(theme={}) {
+export default function getStyle(theme = {}) {
   const appStyle = {...defaultStyle, ...theme};
   return StyleSheet.create({
     container: {
       paddingLeft: 5,
       paddingRight: 5,
       backgroundColor: appStyle.calendarBackground
+    },
+    dayContainer: {
+      flex: 1,
+      alignItems: 'center'
+    },
+    emptyDayContainer: {
+      flex: 1
     },
     monthView: {
       backgroundColor: appStyle.calendarBackground
@@ -23,4 +30,3 @@ export default function getStyle(theme={}) {
     ...(theme[STYLESHEET_ID] || {})
   });
 }
-
