@@ -1,4 +1,10 @@
-export { default as Calendar } from './calendar';
-export { default as CalendarList } from './calendar-list';
-export { default as Agenda } from './agenda';
-export { default as LocaleConfig } from 'xdate';
+export {default as Calendar} from './calendar';
+export {default as CalendarList} from './calendar-list';
+export {default as Agenda} from './agenda';
+export {default as ExpandableCalendar} from './expandableCalendar';
+export {default as AgendaList} from './expandableCalendar/agendaList';
+export {default as Timeline} from './timeline/Timeline';
+export {default as CalendarProvider} from './expandableCalendar/calendarProvider';
+export {default as asCalendarConsumer} from './expandableCalendar/asCalendarConsumer';
+export {default as WeekCalendar} from './expandableCalendar/weekCalendar';
+export {default as LocaleConfig} from 'xdate';
